backend/requirements.txt
Click==7.0
Flask==1.1.2
Flask-Cors==3.0.8
Flask-gzip==0.2
flask-marshmallow==0.11.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
GitPython==3.1.2
marshmallow==3.5.1
marshmallow-sqlalchemy==0.22.3
oyaml==0.9
pandas==1.0.3
python-box==4.2.2
requests==2.23.0
tabulate==0.8.7
xlwt==1.3.0

